+++ backend/common/watchdogs/group_vdbs.py
@@ -9,6 +9,13 @@ class GroupVDBSTable:
         pass
 
     @classmethod
+    async def async_select_by_group_id_order_by_seqn(cls, session: AsyncSession, group_id: int) -> list[GroupVDBs]:
+        where_clause = (GroupVDBs.deleted == 0) & (GroupVDBs.group_id == group_id)
+        stmt = select(GroupVDBs).where(where_clause).order_by(GroupVDBs.gvdbs_seqn)
+        result = await session.execute(stmt)
+        return list(result.scalars().all())
+
+    @classmethod
     async def async_select_all_order_by_group_id_seqn(cls, session: AsyncSession) -> list[GroupVDBs]:
         stmt = select(GroupVDBs).where(GroupVDBs.deleted==0).order_by(GroupVDBs.group_id, GroupVDBs.gvdbs_seqn)
         result = await session.execute(stmt)