+++ backend/cwa_lib/routers/misc.py
@@ -5,6 +5,7 @@ from common.sql_db_async import async_get_session
 from common.sql_models import User
 from cwa_lib.pydantic_schemas.user import ChangePasswordRequest
 from cwa_lib.app import current_active_user, password_helper
+from cwa_lib.pages.server_status import ServerStatusPage
 
 
 router__misc = APIRouter()
@@ -36,4 +37,4 @@ async def server_status(
     user: User = Depends(current_active_user),
     session: AsyncSession = Depends(async_get_session),
 ):
-    return {}
+    return await ServerStatusPage(session, user.group_id).get_all_data()